requirements.txt
# Created automatically by make update-requirements-txt. Do not update manually!
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
decorator==4.4.2
distlib==0.3.6
entrypoints==0.4
fastapi==0.92.0
filelock==3.9.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
identify==2.5.18
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
moviepy==1.0.3
multidict==6.0.4
nodeenv==1.7.0
numpy==1.24.2
openai==0.26.5
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
platformdirs==3.0.0
pre-commit==3.1.0
proglog==0.1.10
protobuf==3.20.3
pyarrow==11.0.0
pydantic==1.10.5
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rich==13.3.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.25.0
streamlit==1.19.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
validators==0.20.0
virtualenv==20.19.0
watchdog==2.3.0
yarl==1.8.2
zipp==3.15.0
